Move authentication to Layout component
